chore: reorganize repo into frontend and backend directories
