+++ frontend/ConsolidationScreen.js
