+++ frontend/HapticTab.tsx
